+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Redesigned mod priority window
 - Vanilla/default mods can now be prioritized over extended material mods
+- Mods can now be disabled such that assets from them are not used for patching your load order
 - "Use MO2 Loose File Order" option has been disabled in favor of the new system
 
 ## [0.8.13] - 2025-07-04